Fix transactions with float amounts

+++ server/backend/src/Queue/Transaction/consumer/processor.ts
@@ -62,7 +62,7 @@ const sendPayoutRequest = async (transactionRecord: PaymentsTransactionRecord, f
   const payoutRequestBody: PayoutRequest = {
     account_number: transactionRecord.source_account,
     // Converting rupees to paisa
-    amount: parseInt(transactionRecord.amount!) * 100,
+    amount: parseFloat(transactionRecord.amount!) * 100,
     currency: transactionRecord.currency,
     fund_account_id: fundId,
     mode: transactionRecord.mode,
